src/requirements.txt
django
djangorestframework
django-rest-enumfield
drf-spectacular
APScheduler